refactor(api): split controllers by feature
