--- apps/api-gateway/src/api/api.controller.ts
@@ -1,328 +1,0 @@
-import { Body, Controller, Get, Param, Post, Put } from '@nestjs/common';
-import { generateObject, generateText } from 'ai';
-import { mailTool } from '../tools/mail.tool';
-import { getAIModel } from '../aiProvider';
-import { PrismaService } from '../prisma.service';
-import { mistral } from '@ai-sdk/mistral';
-import { z } from 'zod';
-
-const stepSchema = z.object({
-  prompt: z.string(),
-  dependencies: z.array(z.number()),
-  idx: z.number().optional(),
-});
-const chatSchema = z.object({
-  title: z.string(),
-  description: z.string(),
-  globalPrompt: z.string().optional(),
-  steps: z.array(stepSchema),
-});
-const universalSchema = z.object({
-  input: z.string().optional(),
-  prompt: z.string(),
-  globalPrompt: z.string().optional(),
-  media: z.string().optional(),
-});
-const fieldSchema = z.object({
-  label: z.string(),
-  name: z.string(),
-  type: z.string(),
-  placeholder: z.string().optional(),
-  options: z.array(z.string()).optional(),
-  description: z.string().optional(),
-  required: z.boolean().optional(),
-});
-const formSchema = z.object({ fields: z.array(fieldSchema) });
-const recordSchema = z.record(z.string(), z.string());
-const extractionServiceSchema = z.object({
-  title: z.string(),
-  schema: z.array(fieldSchema),
-  chatSteps: z.array(stepSchema).optional(),
-  chatGlobalPrompt: z.string().optional(),
-});
-
-@Controller()
-export class ApiController {
-  constructor(private prisma: PrismaService) {}
-
-  @Get('chat-infinite')
-  listChats() {
-    return this.prisma.chatInfinite.findMany({
-      include: { steps: true },
-      orderBy: { createdAt: 'desc' },
-    });
-  }
-
-  @Get('chat-infinite/:id')
-  getChat(@Param('id') id: string) {
-    return this.prisma.chatInfinite.findUnique({
-      where: { id: Number(id) },
-      include: { steps: true },
-    });
-  }
-
-  @Post('chat-infinite')
-  async createChat(@Body() body: unknown) {
-    const { title, description, globalPrompt, steps } = chatSchema.parse(body);
-    return this.prisma.chatInfinite.create({
-      data: {
-        title,
-        description,
-        globalPrompt,
-        steps: {
-          create: steps.map((s, idx) => ({
-            prompt: s.prompt,
-            dependencies: s.dependencies,
-            idx,
-          })),
-        },
-      },
-      include: { steps: true },
-    });
-  }
-
-  @Put('chat-infinite/:id')
-  async updateChat(@Param('id') id: string, @Body() body: unknown) {
-    const { title, description, globalPrompt, steps } = chatSchema.parse(body);
-    await this.prisma.step.deleteMany({ where: { chatId: Number(id) } });
-    return this.prisma.chatInfinite.update({
-      where: { id: Number(id) },
-      data: {
-        title,
-        description,
-        globalPrompt,
-        steps: {
-          create: steps.map((s, idx) => ({
-            prompt: s.prompt,
-            dependencies: s.dependencies,
-            idx,
-          })),
-        },
-      },
-      include: { steps: true },
-    });
-  }
-
-  @Get('extraction-service')
-  listExtractionServices() {
-    return this.prisma.extractionService.findMany({
-      orderBy: { createdAt: 'desc' },
-    });
-  }
-
-  @Get('extraction-service/:id')
-  getExtractionService(@Param('id') id: string) {
-    return this.prisma.extractionService.findUnique({
-      where: { id: Number(id) },
-    });
-  }
-
-  @Post('extraction-service')
-  createExtractionService(@Body() body: unknown) {
-    const { title, schema, chatSteps = [], chatGlobalPrompt = '' } =
-      extractionServiceSchema.parse(body);
-    return this.prisma.extractionService.create({
-      data: { title, schema, chatSteps, chatGlobalPrompt },
-    });
-  }
-
-  @Put('extraction-service/:id')
-  updateExtractionService(@Param('id') id: string, @Body() body: unknown) {
-    const { title, schema, chatSteps = [], chatGlobalPrompt = '' } =
-      extractionServiceSchema.parse(body);
-    return this.prisma.extractionService.update({
-      where: { id: Number(id) },
-      data: { title, schema, chatSteps, chatGlobalPrompt },
-    });
-  }
-
-  @Post('universal-chat')
-  async universalChat(@Body() body: unknown) {
-    const {
-      input = '',
-      prompt,
-      globalPrompt = '',
-      media,
-    } = universalSchema.parse(body);
-    const content: (
-      | { type: 'text'; text: string }
-      | { type: 'image'; image: string }
-    )[] = [{ type: 'text', text: `${prompt}\n${input}`.trim() }];
-    if (media) content.push({ type: 'image', image: media });
-    const messages: { role: 'system' | 'user'; content: any }[] =
-      globalPrompt.trim()
-        ? [
-            { role: 'system', content: globalPrompt },
-            { role: 'user', content },
-          ]
-        : [{ role: 'user', content }];
-    const result = await generateText({
-      model: getAIModel(),
-      messages,
-      tools: { mail: mailTool },
-      maxSteps: 5,
-    });
-    return { output: result.text };
-  }
-
-  @Post('generative-form')
-  async generativeForm(@Body() body: unknown) {
-    const { input } = body as any as { input?: string };
-    const { object } = await generateObject({
-      model: getAIModel(),
-      mode: 'json',
-      schema: formSchema,
-      prompt: `Générez le JSON du formulaire pour "${input}".`,
-      system:
-        `Vous êtes un générateur de formulaires dynamiques pour une plateforme de services digitale en ligne.\nGénérez **uniquement** un objet JSON correspondant à la demande :\n"${input}"\n– Utilisez strictement les types HTML5 pertinents.\n– Pas de champs inutiles, redondants, button, hidden ou password.\n– Les champs doivent être UX-friendly et ordonnés du plus essentiel au plus accessoire.\n- Pas de doublons dans les noms de champs et les options.`.trim(),
-      temperature: 0.2,
-    });
-    console.log('Generated form schema:', object);
-    console.log('input:', input);
-    return object;
-  }
-
-  @Post('extractor')
-  async extractor(@Body() body: unknown) {
-    const { text, image } = body as any as { text?: string; image?: string };
-    if (!image) {
-      return { error: 'Une image ou un PDF est requis' };
-    }
-    const messages: any[] = [
-      {
-        role: 'user',
-        content: [
-          {
-            type: 'text',
-            text: `Extrait le JSON du formulaire à partir de l'image/PDF fourni${text ? ` pour "${text}"` : ''}.`,
-          },
-          { type: 'image', image },
-        ],
-      },
-    ];
-    const { object } = await generateObject({
-      model: mistral('pixtral-large-latest'),
-      system:
-        `Vous êtes un extracteur de formulaires dynamiques pour une plateforme de services digitale en ligne.\nAnalysez l'image/PDF fourni et générez **uniquement** un objet JSON correspondant au formulaire visible.\n– Utilisez strictement les types HTML5 pertinents.\n– Pas de champs inutiles, redondants, button, hidden ou password.\n– Les champs doivent être UX-friendly et ordonnés du plus essentiel au plus accessoire.\n- Pas de doublons dans les noms de champs et les options.\n- Extrayez fidèlement les informations visibles dans l'image/PDF ainsi que ses valeurs.`.trim(),
-      messages,
-      schema: formSchema,
-    });
-    return object;
-  }
-
-  @Post('extract-data')
-  async extractData(@Body() body: unknown) {
-    const data = body as any;
-    if (Array.isArray(data.images)) {
-      const req = z
-        .object({
-          templateId: z.number().optional(),
-          images: z.array(z.string()),
-          fields: z.array(fieldSchema).optional(),
-          chatSteps: z.array(stepSchema).optional(),
-          chatGlobalPrompt: z.string().optional(),
-        })
-        .parse(data);
-
-      let fields = req.fields ?? [];
-      let chatSteps = req.chatSteps ?? [];
-      let chatGlobalPrompt = req.chatGlobalPrompt ?? '';
-      if (req.templateId) {
-        const tpl = await this.prisma.extractionService.findUnique({
-          where: { id: req.templateId },
-        });
-        if (!tpl) return { error: 'template not found' };
-        fields = tpl.schema as any;
-        chatSteps = (tpl.chatSteps as any) ?? [];
-        chatGlobalPrompt = tpl.chatGlobalPrompt ?? '';
-      }
-      if (fields.length === 0) return { error: 'fields missing' };
-      const records: Record<string, string>[] = [];
-      for (const image of req.images) {
-        const description = fields
-          .map((f) => `${f.label} (${f.type}) -> ${f.name}`)
-          .join('; ');
-        const messages: any[] = [
-          {
-            role: 'user',
-            content: [
-              {
-                type: 'text',
-                text: `Extrait uniquement les valeurs des champs suivants depuis l'image fournie. Rends un objet JSON avec les noms techniques : ${description}`,
-              },
-              { type: 'image', image },
-            ],
-          },
-        ];
-        const { object } = await generateObject({
-          model: mistral('pixtral-large-latest'),
-          system:
-            `Vous êtes un service d'extraction de données. Vous recevez une image et la liste des champs attendus. Retournez uniquement un objet JSON avec les valeurs extraites.`.trim(),
-          messages,
-          schema: recordSchema,
-        });
-        records.push(object);
-      }
-      const outputs: string[] = [];
-      const inputBase = JSON.stringify(records);
-      const steps = [...chatSteps].sort((a, b) => (a.idx ?? 0) - (b.idx ?? 0));
-      for (let i = 0; i < steps.length; i++) {
-        const step = steps[i];
-        if (!step.prompt.trim()) {
-          outputs[i] = '';
-          continue;
-        }
-        const deps = step.dependencies.length > 0 ? step.dependencies : [i - 1];
-        const input = deps
-          .map((d) => (d === -1 ? inputBase : outputs[d] || ''))
-          .join('\n');
-        const messages: { role: 'system' | 'user'; content: any }[] =
-          chatGlobalPrompt.trim()
-            ? [
-                { role: 'system', content: chatGlobalPrompt },
-                { role: 'user', content: `${step.prompt}\n${input}`.trim() },
-              ]
-            : [{ role: 'user', content: `${step.prompt}\n${input}`.trim() }];
-        const result = await generateText({
-          model: getAIModel(),
-          messages,
-          tools: { mail: mailTool },
-          maxSteps: 5,
-        });
-        outputs[i] = result.text;
-      }
-      return { records, outputs };
-    }
-
-    const { fields, image } = data as {
-      fields?: { name: string; label: string; type: string }[];
-      image?: string;
-    };
-    if (!image || !fields) {
-      return { error: 'Image et champs requis' };
-    }
-    const description = fields
-      .map((f) => `${f.label} (${f.type}) -> ${f.name}`)
-      .join('; ');
-    const messages: any[] = [
-      {
-        role: 'user',
-        content: [
-          {
-            type: 'text',
-            text: `Extrait uniquement les valeurs des champs suivants depuis l'image fournie. Rends un objet JSON avec les noms techniques : ${description}`,
-          },
-          { type: 'image', image },
-        ],
-      },
-    ];
-    const { object } = await generateObject({
-      model: mistral('pixtral-large-latest'),
-      system:
-        `Vous êtes un service d'extraction de données. Vous recevez une image et la liste des champs attendus. Retournez uniquement un objet JSON avec les valeurs extraites.`.trim(),
-      messages,
-      schema: recordSchema,
-    });
-    return object;
-  }
-}
