+++ apps/api-gateway/src/api/api.module.ts
@@ -1,9 +1,19 @@
 import { Module } from '@nestjs/common';
-import { ApiController } from './api.controller';
+import { ChatInfiniteController } from './chat-infinite/chat-infinite.controller';
+import { ExtractionServiceController } from './extraction-service/extraction-service.controller';
+import { UniversalChatController } from './universal-chat/universal-chat.controller';
+import { ExtractorController } from './extractor/extractor.controller';
+import { GenerativeFormController } from './generative-form/generative-form.controller';
 import { PrismaService } from '../prisma.service';
 
 @Module({
-  controllers: [ApiController],
+  controllers: [
+    ChatInfiniteController,
+    ExtractionServiceController,
+    UniversalChatController,
+    ExtractorController,
+    GenerativeFormController,
+  ],
   providers: [PrismaService],
 })
 export class ApiModule {}
